requirements.txt
pandas
python-dotenv
openai
openpyxl
statsmodels
pingouin